231206 - 명령어 한국어로  추가 / tle 세부 분리

+++ main.py
@@ -77,10 +77,15 @@ async def on_ready():
 # 메세지 전송 이벤트 처리    
 @app.event
 async def on_message(ctx : discord.Message) -> None:
+  
     if ctx.channel.id == log_channel or ctx.channel.id == feedback_log_channel:
-        not_here_embed = discord.Embed(title= '여기는 로그가 남는 채널입니다', description= f'5초 후에 지워집니다.', colour= 0xfe7866)
-        await ctx.delete()
-        await ctx.channel.send(embed= not_here_embed, delete_after= 5, mention_author= True)
+        if ctx.author.bot == True:
+            pass
+        
+        else:
+            not_here_embed = discord.Embed(title= '여기는 로그가 남는 채널입니다', description= f'5초 후에 지워집니다.', colour= 0xfe7866)
+            await ctx.delete()
+            await ctx.channel.send(embed= not_here_embed, delete_after= 5, mention_author= True)
 
     else:
         pass
